Move all medium components to one page and reactive with boolean value for difficulty
